refactor : S3Servie 인터페이스로 분리 후 구현체 이름 변경 #20
