+++ src/test/java/com/usememo/jugger/global/s3/service/S3ServiceImplementationTest.java
@@ -3,7 +3,6 @@ package com.usememo.jugger.global.s3.service;
 import io.awspring.cloud.s3.S3Template;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.mockito.Mockito;
 import org.springframework.core.io.buffer.DataBuffer;
 import org.springframework.core.io.buffer.DefaultDataBufferFactory;
 import org.springframework.http.codec.multipart.FilePart;
@@ -16,14 +15,14 @@ import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.*;
 
-class S3ServiceTest {
+class S3ServiceImplementationTest {
 
 	@Test
 	@DisplayName("S3 업로드 테스트")
 	void uploadFile_shouldReturnUrl_whenSuccessful() {
 		// given
 		S3Template s3Template = mock(S3Template.class);
-		S3Service s3Service = new S3Service(s3Template);
+		S3ServiceImplementation s3ServiceImplementation = new S3ServiceImplementation(s3Template);
 
 		FilePart filePart = mock(FilePart.class);
 		when(filePart.filename()).thenReturn("test-image.jpg");
@@ -35,7 +34,7 @@ class S3ServiceTest {
 			.thenReturn(null);
 
 		// when & then
-		StepVerifier.create(s3Service.uploadFile(filePart))
+		StepVerifier.create(s3ServiceImplementation.uploadFile(filePart))
 			.assertNext(url -> {
 				assert url.startsWith("https://jugger-bucket.s3.ap-northeast-2.amazonaws.com/");
 				assert url.endsWith(".jpg");
